+++ app/src/main/java/ru/practicum/android/diploma/data/dto/NotFoundException.kt
@@ -1,0 +1,3 @@
+package ru.practicum.android.diploma.data.dto
+
+class NotFoundException(message: String) : Exception(message)
