+++ app/src/main/java/ru/practicum/android/diploma/data/vacancy/VacancyRepository.kt
@@ -3,29 +3,32 @@ package ru.practicum.android.diploma.data.vacancy
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import retrofit2.HttpException
+import ru.practicum.android.diploma.R
 import ru.practicum.android.diploma.data.dto.NetworkResponse
+import ru.practicum.android.diploma.data.dto.NotFoundException
 import ru.practicum.android.diploma.data.dto.VacancyDetails
 import ru.practicum.android.diploma.data.mappers.VacancyResponseToDomainMapper
 import ru.practicum.android.diploma.data.network.HeadHunterNetworkClient
+import ru.practicum.android.diploma.domain.search.models.DomainVacancy
 import ru.practicum.android.diploma.util.SERVER_ERROR
 import ru.practicum.android.diploma.util.SUCCESS_CODE
-import ru.practicum.android.diploma.domain.search.models.DomainVacancy
 import java.io.IOException
 
 class VacancyRepository(
     private val networkClient: HeadHunterNetworkClient,
-    private val vacancyMapper: VacancyResponseToDomainMapper
+    private val vacancyMapper: VacancyResponseToDomainMapper,
 ) {
     suspend fun getVacancyDetails(id: String): NetworkResponse<DomainVacancy> {
         return withContext(Dispatchers.IO) {
             try {
                 val response = networkClient.getVacancy(id)
                 if (response.isSuccessful) {
-                    val vacancyDetails: VacancyDetails = response.body() ?: throw NotFoundException("Vacancy not found")
+                    val vacancyDetails: VacancyDetails =
+                        response.body() ?: throw NotFoundException(R.string.server_error.toString())
                     val domainVacancyDetails = vacancyMapper.map(vacancyDetails, true)
                     NetworkResponse(domainVacancyDetails, SUCCESS_CODE)
                 } else {
-                    val errorBody = response.errorBody()?.string() ?: "Unknown error"
+                    val errorBody = response.errorBody()?.string() ?: R.string.server_error.toString()
                     NetworkResponse(null, SERVER_ERROR, errorBody)
                 }
             } catch (ex: IOException) {
@@ -36,5 +39,3 @@ class VacancyRepository(
         }
     }
 }
-
-class NotFoundException(message: String) : Exception(message)
