+++ app/src/main/java/ru/practicum/android/diploma/util/Resource.kt
@@ -2,5 +2,5 @@ package ru.practicum.android.diploma.util
 
 sealed class Resource<T>(val data: T? = null, val resultCode: Int? = null) {
     class Success<T>(data: T?) : Resource<T>(data)
-    class Error<T>(resultCode: Int, data: T? = null) : Resource<T>(data, resultCode)
+    class Error<T>(resultCode: Int) : Resource<T>(resultCode = resultCode)
 }
