+++ app/src/main/res/drawable/bottom_nav_selector.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_checked="true" android:color="?attr/colorPrimary"/>
+    <item android:state_checked="false" android:color="@color/gray"/>
+</selector>
