+++ app/src/main/res/layout/company_info.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    xmlns:tools="http://schemas.android.com/tools"
     android:background="@drawable/bg_company_info"
     android:orientation="horizontal"
     android:padding="16dp"
@@ -13,9 +13,9 @@
         android:layout_width="50dp"
         android:layout_height="50dp"
         android:layout_marginEnd="8dp"
-        android:src="@drawable/placeholder_android_xml_24"
         android:background="?attr/selectableItemBackgroundBorderless"
-        android:scaleType="centerCrop" />
+        android:scaleType="centerCrop"
+        android:src="@drawable/placeholder_android_xml_24" />
 
     <LinearLayout
         android:layout_width="0dp"
@@ -24,13 +24,15 @@
         android:orientation="vertical">
 
         <TextView
-            style="@style/vacancy_title_style"
             android:id="@+id/companyName"
-            tools:text="Еда"/>
+            style="@style/vacancy_title_style"
+            android:textColor="@color/black"
+            tools:text="Еда" />
 
         <TextView
-            style="@style/vacancy_text_style"
             android:id="@+id/companyLocation"
+            style="@style/vacancy_text_style"
+            android:textColor="@color/black"
             tools:text="Москва" />
     </LinearLayout>
 </LinearLayout>
