+++ app/src/main/res/values-night/themes.xml
@@ -12,10 +12,7 @@
         <!-- Status bar color. -->
         <item name="android:statusBarColor">?attr/colorSecondary</item>
         <!-- Customize your dark theme here. -->
-        <!-- <item name="colorPrimary">@color/my_dark_primary</item> -->
         <item name="android:windowBackground">@color/black</item>
         <item name="android:windowLightStatusBar">false</item>
-        <item name="textInVacancyDatails">@color/black</item>
     </style>
-
 </resources>
