--- app/src/main/res/values/attrs.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <attr name="textInVacancyDatails" format="color"/>
-</resources>
