+++ app/src/main/res/values/themes.xml
@@ -12,15 +12,12 @@
         <!-- Status bar color. -->
         <item name="android:statusBarColor">?attr/colorSecondary</item>
         <!-- Customize your light theme here. -->
-        <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
         <item name="android:windowBackground">@color/white</item>
         <item name="android:windowLightStatusBar">true</item>
-        <item name="textInVacancyDatails">@color/black</item>
         <item name="radioButtonStyle">@style/Widget.App.RadioButton</item>
     </style>
 
     <style name="Widget.App.RadioButton" parent="Widget.MaterialComponents.CompoundButton.RadioButton">
         <item name="buttonTint">@color/blue</item>
     </style>
-
 </resources>
